Move GlyphClass into own module
