+++ fea-rs/src/common.rs
@@ -1,27 +1,18 @@
 //! Types and helpers shared across modules
 
-use std::{
-    fmt::{Display, Formatter},
-    rc::Rc,
-};
+use std::fmt::{Display, Formatter};
 
 use smol_str::SmolStr;
 pub use write_fonts::types::GlyphId;
 
+mod glyph_class;
 mod glyph_map;
 
+pub use glyph_class::GlyphClass;
 pub use glyph_map::GlyphMap;
 
 /// A glyph name
 pub type GlyphName = SmolStr;
-
-/// A glyph class, as used in the FEA spec.
-///
-/// This type is currently somewhat confused; in certain places the spec expects
-/// that a glyph class is sorted and deduplicated, and in other places it expects
-/// a glyph class to be an arbitrary sequence of glyphs.
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct GlyphClass(Rc<[GlyphId]>);
 
 /// A glyph or glyph class.
 ///
@@ -43,32 +34,6 @@ pub enum GlyphIdent {
     Name(GlyphName),
     /// a CID
     Cid(u16),
-}
-
-impl From<GlyphOrClass> for GlyphClass {
-    fn from(src: GlyphOrClass) -> GlyphClass {
-        match src {
-            GlyphOrClass::Class(class) => class,
-            GlyphOrClass::Glyph(id) => id.into(),
-            GlyphOrClass::Null => GlyphClass(Rc::new([])),
-        }
-    }
-}
-
-impl std::iter::FromIterator<GlyphId> for GlyphClass {
-    fn from_iter<T: IntoIterator<Item = GlyphId>>(iter: T) -> Self {
-        GlyphClass(iter.into_iter().collect())
-    }
-}
-
-impl<'a> std::iter::IntoIterator for &'a GlyphClass {
-    type Item = &'a GlyphId;
-
-    type IntoIter = std::slice::Iter<'a, GlyphId>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.0.iter()
-    }
 }
 
 impl<T: Into<GlyphName>> From<T> for GlyphIdent {
@@ -122,7 +87,7 @@ impl GlyphOrClass {
         std::iter::from_fn(move || {
             let next = match &self {
                 GlyphOrClass::Glyph(id) if idx == 0 => Some(*id),
-                GlyphOrClass::Class(cls) => cls.0.get(idx).copied(),
+                GlyphOrClass::Class(cls) => cls.items().get(idx).copied(),
                 _ => None,
             };
             idx += 1;
@@ -140,47 +105,12 @@ impl GlyphOrClass {
             let next = match &self {
                 GlyphOrClass::Glyph(id) if idx == 0 => Some(*id),
                 GlyphOrClass::Null if idx == 0 => Some(GlyphId::NOTDEF),
-                GlyphOrClass::Class(cls) => cls.0.get(idx).copied(),
+                GlyphOrClass::Class(cls) => cls.items().get(idx).copied(),
                 _ => None,
             };
             idx += 1;
             idx %= self.len();
             next
         })
-    }
-}
-
-impl GlyphClass {
-    pub fn items(&self) -> &[GlyphId] {
-        &self.0
-    }
-
-    pub fn sort_and_dedupe(&self) -> GlyphClass {
-        //idfk I guess this is fine
-        let mut vec = self.0.iter().cloned().collect::<Vec<_>>();
-        vec.sort_unstable();
-        vec.dedup();
-        GlyphClass(vec.into())
-    }
-
-    pub fn iter(&self) -> impl Iterator<Item = GlyphId> + '_ {
-        self.items().iter().copied()
-    }
-
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-}
-
-impl From<Vec<GlyphId>> for GlyphClass {
-    fn from(src: Vec<GlyphId>) -> GlyphClass {
-        GlyphClass(src.into())
-    }
-}
-
-impl From<GlyphId> for GlyphClass {
-    fn from(src: GlyphId) -> GlyphClass {
-        let slice: &[_] = &[src];
-        GlyphClass(slice.into())
     }
 }
